+++ frontend/tests/unit/components/EntryFeed.spec.js
@@ -1,0 +1,49 @@
+import {shallowMount, createLocalVue} from '@vue/test-utils';
+
+import BootstrapVue, {BButton} from 'bootstrap-vue';
+
+import EntryFeed from '@/components/EntryFeed.vue';
+import PartialJournal from '@/components/PartialJournal.vue';
+
+const localVue = createLocalVue();
+localVue.use(BootstrapVue);
+
+describe('EntryFeed component', () => {
+  test('renders entries as PartialJournals', () => {
+    const mockEntries = [
+      {
+        key: '/testUser321/2019-09-27',
+        author: 'testUser321',
+        date: new Date(2019, 9, 27),
+        markdown: 'I went to the zoo today',
+      },
+      {
+        key: '/testUser456/2019-09-27',
+        author: 'testUser456',
+        date: new Date(2019, 9, 27),
+        markdown: 'I ate an ice-cream sandwich yesterday',
+      },
+    ];
+
+    const wrapper = shallowMount(EntryFeed, {
+      propsData: {
+        readEntriesFromStore: () => {
+          return mockEntries;
+        },
+        readEntriesFromServer: () => {
+          return new Promise(function(resolve) {
+            resolve(mockEntries);
+          });
+        },
+      },
+      localVue,
+    });
+
+    const entries = wrapper.findAll(PartialJournal);
+    expect(entries).toHaveLength(mockEntries.length);
+
+    entries.wrappers.forEach((wrapper, i) => {
+      expect(wrapper.vm.entry).toBe(mockEntries[i]);
+    });
+  });
+});
